+++ primefaces/src/main/java/org/primefaces/util/LangUtils.java
@@ -27,19 +27,15 @@ import java.beans.BeanInfo;
 import java.beans.IntrospectionException;
 import java.beans.Introspector;
 import java.beans.PropertyDescriptor;
-import java.lang.reflect.*;
+import java.lang.reflect.Method;
+import java.lang.reflect.ParameterizedType;
+import java.lang.reflect.Type;
+import java.lang.reflect.TypeVariable;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.*;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 import java.util.regex.Pattern;
-import javax.el.ELException;
 import javax.faces.FacesException;
-import javax.faces.component.UIComponent;
-import javax.faces.context.FacesContext;
-import javax.faces.convert.Converter;
-import javax.faces.convert.ConverterException;
 import javax.xml.bind.DatatypeConverter;
 
 public class LangUtils {
@@ -511,57 +507,5 @@ public class LangUtils {
             }
         }
         return true;
-    }
-
-    public static Object convertToType(Object value, Class<?> valueType, Class<?> logContext) {
-        // skip null
-        if (value == null) {
-            return null;
-        }
-
-        // its already the same type
-        if (valueType.isAssignableFrom(value.getClass())) {
-            return value;
-        }
-
-        FacesContext context = FacesContext.getCurrentInstance();
-        Logger logger = Logger.getLogger(logContext.getName());
-
-        // primivites dont need complex conversion, try the ELContext first
-        if (BeanUtils.isPrimitiveOrPrimitiveWrapper(valueType)) {
-            try {
-                return context.getELContext().convertToType(value, valueType);
-            }
-            catch (ELException e) {
-                logger.log(Level.INFO, e, () -> "Could not convert '" + value + "' to " + valueType + " via ELContext!");
-            }
-        }
-
-        Converter targetConverter = context.getApplication().createConverter(valueType);
-        if (targetConverter == null) {
-            logger.log(Level.FINE, () -> "Skip conversion as no converter was found for " + valueType
-                    + "; Create a JSF Converter for it or overwrite Object convertToType(String value, Class<?> valueType)!");
-            return value;
-        }
-
-        Converter sourceConverter = context.getApplication().createConverter(value.getClass());
-        if (sourceConverter == null) {
-            logger.log(Level.FINE, () -> "Skip conversion as no converter was found for " + value.getClass()
-                    + "; Create a JSF Converter for it or overwrite Object convertToType(String value, Class<?> valueType)!");
-        }
-
-        // first convert the object to string
-        String stringValue = sourceConverter == null
-                ? value.toString()
-                : sourceConverter.getAsString(context, UIComponent.getCurrentComponent(context), value);
-
-        // now convert the string to the required target
-        try {
-            return targetConverter.getAsObject(context, UIComponent.getCurrentComponent(context), stringValue);
-        }
-        catch (ConverterException e) {
-            logger.log(Level.INFO, e, () -> "Could not convert '" + stringValue + "' to " + valueType + " via " + targetConverter.getClass().getName());
-            return value;
-        }
     }
 }
